requirements.txt
# Requirements for Python scripts in this repo that have automation enabled!

absl-py
arcgis2geojson
dataclasses==0.6
datacommons==1.4.3
frozendict==1.2
func-timeout==4.3.5
geojson==2.5.0
google-cloud-bigquery
google-cloud-storage>=2.7.0
google-cloud-logging==3.4.0
google-cloud-scheduler==2.10.0
gspread
lxml==4.9.1
numpy
pandas
pylint
pytest
requests==2.27.1
retry==0.9.2
shapely==1.8.5
urllib3==1.26.8
xarray==0.19.0
xlrd==1.2.0
zipp
beautifulsoup4
ratelimit
